move database connection config out of server/config.ts
kysely-ctl loads its config through jiti, which has its own module resolver
and ignores Node's --conditions flag. So even with NODE_OPTIONS=--conditions=react-server
on the npm script, the migration entry point hits the server-only throw via
dialect.ts → server/config.ts.

Break the chain: introduce src/database/env.ts with databaseURL and DEPLOYED_ENV.
dialect.ts imports from there, server/config.ts re-exports databaseURL so existing
server-side callers keep working.
